File restructuring
Reorganized files to follow the recent file path pattern
