+++ library-window/logbook/logbook.cpp
